I have a use case where I stich the request object with the corresponding error message. I don't see any direct example which merges two schema -one of the request and another of response and a simple write(req, res) should give the csv.
So here I am converting request to JSON and then adding response to this Json object. collecting all such json object in a Json array and using various other libraries available to get the csv string and directly writing to the response writer.

There is no solution to preserve the order of headers. But this will do for my usecase